say that a clamped component fits in a byte
    ImageConvertor.h(252): warning C4838: conversion from 'uint32_t' to
    'uint8_t' requires a narrowing conversion

Three times, once per channel: the CVec4 to SColor24 conversion builds the
struct with a braced initialiser, MakeComponent returns uint32_t, and narrowing
inside braces is ill-formed however small the value turns out to be.

The value is always small. MakeComponent is

    Clamp( f * 255.0f, 0.0f, 255.0f )

so it cannot leave 0..255, and the conversion the compiler was objecting to
never lost anything. The casts state what the clamp already guarantees, and the
generated code is the same.

MakeComponent keeps its uint32_t return type, which is the other way this could
have been fixed. Its remaining caller builds a packed ARGB word and shifts the
first component left by 24; as a uint8_t that would promote to int, and any
channel above 127 would overflow a signed 32 bit shift. Narrowing once, where
the target really is a byte, is the smaller change and the safer one.

SColor24 is b, g, r in that order and the initialiser reads z, y, x, which is
B, G, R. That pairing is unchanged and is the same one the uint32_t overload two
lines up uses.

Image now compiles with no warnings at all, and its 19 unit tests pass.

+++ Versions/Temporary/Engine/Sources/Image/ImageConvertor.h
@@ -249,7 +249,22 @@ public:
 	CRawColorConvertor( int nColorMapLength, int nColorMapEntrySize, CDataStream *pStream ) {}
 	SColor24 operator()( const uint32_t input ) const { const SColor24 color = { GetB(input), GetG(input), GetR(input)	}; return color; }
 	SColor24 operator()( const SColor24 input ) const { return input; }
-	SColor24 operator()( const CVec4 &input ) const { const SColor24 color = { MakeComponent(input.z), MakeComponent(input.y), MakeComponent(input.x) }; return color; }
+	// MakeComponent clamps to 0..255 and so always fits a uint8_t, but a braced
+	// initialiser does not take that on trust: narrowing inside one is
+	// ill-formed, and MSVC says so once per channel (C4838). The casts state
+	// what the clamp already guarantees.
+	//
+	// MakeComponent keeps returning uint32_t rather than uint8_t, which would
+	// have removed the need for them: its other caller shifts the result left by
+	// 24, and a uint8_t there would promote to int and overflow on any channel
+	// above 127.
+	SColor24 operator()( const CVec4 &input ) const
+	{
+		const SColor24 color = { static_cast<uint8_t>( MakeComponent(input.z) ),
+		                         static_cast<uint8_t>( MakeComponent(input.y) ),
+		                         static_cast<uint8_t>( MakeComponent(input.x) ) };
+		return color;
+	}
 	bool IsReady() const { return true; }
 };
 template <> class CRawColorConvertor<uint8_t>
